+++ app/js/app.js
@@ -3,13 +3,17 @@
 
 // Declare app level module which depends on filters, and services
 var myApp = angular.module('myApp', ['myApp.filters', 'myApp.services', 'myApp.directives', 'ngI18n']).
-  config(['$routeProvider', function($routeProvider) {
-    $routeProvider.when('/view1', {templateUrl: 'partials/partial1.html', controller: MyCtrl1});
-    $routeProvider.when('/view2', {templateUrl: 'partials/partial2.html', controller: MyCtrl2});
-    $routeProvider.otherwise({redirectTo: '/view1'});
-  }]);
+    config(['$routeProvider', function ($routeProvider) {
+    $routeProvider.when('/view1', {templateUrl:'partials/partial1.html', controller:MyCtrl1});
+    $routeProvider.when('/view2', {templateUrl:'partials/partial2.html', controller:MyCtrl2});
+    $routeProvider.otherwise({redirectTo:'/view1'});
+}]);
 
 myApp.value('ngI18nConfig', {
+    //defaultLocale should always be provided!!
+    defaultLocale: 'en',
+    //supportedLocales should always be provided!!
+    supportedLocales: ['en', 'nl'],
     //without leading and trailing slashes
-    basePath: "app/i18n"
+    basePath:'app/i18n'
 });