+++ test/unit/servicesSpec.js
@@ -1,103 +1,139 @@
 'use strict';
 
 describe('ngI18nService', function () {
-    var DEFAULT_LOCALE, DEFAULT_LANGUAGE, DEFAULT_USER_LANGUAGE, LOCALE_WITH_COUNTRY, ngI18nConfig;
+    var ngI18nConfig, windowStub;
     beforeEach(function () {
+
         module('ngI18nConfig');
 
         module(function ($provide) {
-            ngI18nConfig = {};
+            ngI18nConfig = {supportedLocales:['nl', 'en']};
             $provide.value('ngI18nConfig', ngI18nConfig);
+
+            windowStub = {
+                navigator:{
+                    language:undefined,
+                    userLanguage:undefined
+                }
+
+            };
+            $provide.value('$window', windowStub);
         });
 
         module('ngI18nService');
-        DEFAULT_LOCALE = 'nl';
-        DEFAULT_LANGUAGE = 'fr';
-        DEFAULT_USER_LANGUAGE = 'en';
-        LOCALE_WITH_COUNTRY = "nl-be";
+
     });
 
-    describe('ngI18nResourceBundleLoader', function () {
-        var $httpBackend, ngI18nResourceBundleLoader, ngI18nLocaleContextHolder, resourceBundle_nl;
+    describe('ngI18nResourceBundle', function () {
+        var $httpBackend, ngI18nResourceBundle, expectedResourceBundle, resourceBundle;
 
-        beforeEach(inject(function($injector) {
+        beforeEach(inject(function ($injector) {
             $httpBackend = $injector.get('$httpBackend');
-            ngI18nResourceBundleLoader = $injector.get('ngI18nResourceBundleLoader');
-            ngI18nLocaleContextHolder = $injector.get('ngI18nLocaleContextHolder');
-            resourceBundle_nl =   {"key1" : 'waarde1', "key2": 'waarde2'};
+            ngI18nResourceBundle = $injector.get('ngI18nResourceBundle');
+            expectedResourceBundle = {"key1":'value1', "key2":'value2'};
         }));
 
-        afterEach(function() {
+        afterEach(function () {
             $httpBackend.verifyNoOutstandingExpectation();
             $httpBackend.verifyNoOutstandingRequest();
         });
 
-        it("should be able to get resourceBundle for locale from localeContextHolder", function(){
-            var resourceBundle;
-            ngI18nLocaleContextHolder.setLocale(DEFAULT_LOCALE);
-            $httpBackend.when('GET', '/i18n/resourceBundle_nl.json').respond(resourceBundle_nl);
+        it("should be able to get resourceBundle for locale set in options", function () {
+            assertResourceBundleLoaded('nl');
+        });
 
-            ngI18nResourceBundleLoader.get().success(function(data) {
+        describe('Locale from browser', function () {
+            it("should be able to get resourceBundle for Locale from browser (non IE) when locale not set in options", function () {
+                var locale = 'fr';
+                ngI18nConfig.supportedLocales = [locale];
+                windowStub.navigator.language = locale;
+                assertResourceBundleLoadedWithLocaleFromBrowser(locale);
+            });
+
+            it("should be able to get resourceBundle for Locale from browser (IE) when locale not set in options", function () {
+                var locale = 'fr';
+                ngI18nConfig.supportedLocales = [locale];
+                windowStub.navigator.userLanguage = locale;
+                assertResourceBundleLoadedWithLocaleFromBrowser(locale);
+            });
+
+            function assertResourceBundleLoadedWithLocaleFromBrowser(locale) {
+                $httpBackend.when("GET", '/i18n/resourceBundle_' + locale + '.json').respond(expectedResourceBundle);
+
+                ngI18nResourceBundle.get().success(function (data) {
+                    resourceBundle = data;
+                });
+
+                $httpBackend.flush();
+
+                expect(resourceBundle).toEqual(expectedResourceBundle);
+            }
+        });
+
+        describe('base path', function () {
+            it("should be able to configure base path for url", function () {
+                ngI18nConfig.basePath = 'new/base/path';
+                var basePath = 'new/base/path';
+                assertResourceBundleLoaded('nl', basePath);
+            });
+        });
+
+        describe('supported locales', function () {
+            it("should be able to fallback to default resourceBundle when locale not supported", function () {
+                ngI18nConfig.supportedLocales = ['nl', 'en'];
+                assertDefaultResourceBundleLoaded('fr');
+            });
+
+            it("should be able to fallback to resourceBundle for language from Locale when language of locale is supported but locale not", function () {
+                ngI18nConfig.supportedLocales = ['nl', 'en'];
+                assertResourceBundleLoadedForSupportedLocales('nl-be', 'nl');
+            });
+
+            function assertResourceBundleLoadedForSupportedLocales(localeForGet, localeForUrl) {
+                $httpBackend.when("GET", '/i18n/resourceBundle_' + localeForUrl + '.json').respond(expectedResourceBundle);
+
+                ngI18nResourceBundle.get({locale:localeForGet}).success(function (data) {
+                    resourceBundle = data;
+                });
+
+                $httpBackend.flush();
+
+                expect(resourceBundle).toEqual(expectedResourceBundle);
+            }
+        });
+
+        describe('default locale', function () {
+            it('should get default resourceBundle when locale matches default locale', function () {
+                ngI18nConfig.supportedLocales = ['nl', 'en'];
+                ngI18nConfig.defaultLocale = 'nl';
+                assertDefaultResourceBundleLoaded('nl');
+            });
+        });
+
+        function assertResourceBundleLoaded(locale, basePath) {
+            var _basePath = basePath || 'i18n';
+            $httpBackend.when("GET", '/' + _basePath + '/resourceBundle_' + locale + '.json').respond(expectedResourceBundle);
+
+            ngI18nResourceBundle.get({locale:locale}).success(function (data) {
                 resourceBundle = data;
             });
 
             $httpBackend.flush();
 
-            //noinspection JSUnusedAssignment
-            expect(resourceBundle).toEqual(resourceBundle_nl);
+            expect(resourceBundle).toEqual(expectedResourceBundle);
+        }
 
-        });
+        function assertDefaultResourceBundleLoaded(locale) {
+            $httpBackend.when("GET", '/i18n/resourceBundle.json').respond(expectedResourceBundle);
 
-        it("should be able to get configure base path for url", function(){
-            var resourceBundle;
-            ngI18nLocaleContextHolder.setLocale(DEFAULT_LOCALE);
-            ngI18nConfig.basePath = 'new/base/path'
-            $httpBackend.when('GET', '/new/base/path/resourceBundle_nl.json').respond(resourceBundle_nl);
-
-            ngI18nResourceBundleLoader.get().success(function(data) {
+            ngI18nResourceBundle.get({locale:locale}).success(function (data) {
                 resourceBundle = data;
             });
 
             $httpBackend.flush();
 
-            expect(resourceBundle).toEqual(resourceBundle_nl);
-
-        });
-    });
-
-    describe('ngI18nLocaleContextHolder', function () {
-        var ngI18nLocaleContextHolder, windowStub;
-
-        beforeEach(function () {
-            module(function ($provide) {
-                windowStub = {
-                    navigator:{
-                        language:DEFAULT_LANGUAGE,
-                        userLanguage:undefined
-                    }
-
-                };
-                $provide.value('$window', windowStub);
-            });
-            inject(function ($injector) {
-                ngI18nLocaleContextHolder = $injector.get('ngI18nLocaleContextHolder');
-            });
-        });
-
-        it('should able to set Locale', function () {
-            ngI18nLocaleContextHolder.setLocale(DEFAULT_LOCALE);
-            expect(ngI18nLocaleContextHolder.getLocale()).toEqual(DEFAULT_LOCALE);
-
-        });
-
-        it('should able get Locale from browser (non IE) when locale not set', function () {
-            expect(ngI18nLocaleContextHolder.getLocale()).toEqual(DEFAULT_LANGUAGE);
-        });
-
-        it('should able get Locale from browser (IE) when locale not set', function () {
-            windowStub.navigator.userLanguage = DEFAULT_USER_LANGUAGE;
-            expect(ngI18nLocaleContextHolder.getLocale()).toEqual(DEFAULT_USER_LANGUAGE);
-        });
+            expect(resourceBundle).toEqual(expectedResourceBundle);
+        }
     });
 
 });