+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ### v0.2.0
 
 * Added global config basePath
+* Added global config defaultLocale
+* Added global config supportedLocales
+* Renamed ngI18nResourceBundleLoader to ngI18nResourceBundle
+* Removed ngI18nLocaleContextHolder
+* Implemented algorithm to get resourceBundle
 
 ### v0.1.0